MVC Implementado
MVC implementado en el código. Esto quiere decir que el código se ha dividido en cuatro carpetas: modelo, vista, controlador y programa. Siendo programa la que da inicio a la aplicación.
