--- src/mvcSort/Main.java
@@ -1,24 +1,0 @@
-package mvcSort;
-
-//import algoritmos.*;
-//import mvcSort.Aplicacion;
-
-public class Main {
-	public static void main(String [] args) {
-		Aplicacion aplicacion = new Aplicacion();
-		aplicacion.crearVentana("Lista Ordenada");
-		/*
-		Archivo leerArchivo = new Archivo();
-		//leerArchivo.leer("Numeros.txt");
-		leerArchivo.leer("NumerosOrdenados.txt");
-		//leerArchivo.imprimirDatos();
-		Integer [] lista = leerArchivo.convertirArrayEntero();
-		QuickSort quickSort = new QuickSort();
-		quickSort.ordenarLista(lista,1);
-		leerArchivo.escribir("NumerosOrdenados.txt",lista);
-		BusquedaBinaria bB = new BusquedaBinaria();
-		System.out.println(bB.busquedaBinaria(lista, 1530796302, 2));
-		*/
-		
-	}
-}
